+++ src/CoreEx/RefData/ReferenceDataOrchestrator.cs
@@ -16,7 +16,6 @@ using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using System.Transactions;
 
 namespace CoreEx.RefData
 {